+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 - 2021-06-04
+
+### Added
+
+- Added the `:producer` key to each event's `.metadata` map for InOrder
+  subscriptions
+    - This can inform the consumer of how it should store stream positions
+
 ## 0.2.0 - 2021-06-01
 
 ### Changed